Move semisl helpers to utils
